FestivalAdministration und FestivalSuchobjekt

+++ Festiva/src/standardPackage/Artikel.java
@@ -9,7 +9,7 @@ public class Artikel {
 	
 	public int id;
 	public String beschreibung;
-	public double preis;
+	public float preis;
 	public boolean istGelöscht;
 	public int festivalID;
 	
@@ -22,7 +22,7 @@ public class Artikel {
 	 * @param p_istGelöscht: zeigt, ob der Artikel logisch gelöscht wurde
 	 * @param p_festivalID: eindeutige ID des Festivals, zu dem der Artikel gehört
 	 */
-	public Artikel(int p_id, String p_beschreibung, double p_preis, boolean p_istGelöscht, int p_festivalID) {
+	public Artikel(int p_id, String p_beschreibung, float p_preis, boolean p_istGelöscht, int p_festivalID) {
 		this.id = p_id;
 		this.beschreibung = p_beschreibung;
 		this.preis = p_preis;
